+++ HISTORY.rst
@@ -6,7 +6,7 @@ Changelog
 
 * Fix ``Parser.use_kwargs`` behavior when an ``Arg`` is allowed missing. The ``allow_missing`` attribute is ignored when ``use_kwargs`` is called.
 * Allow ``ValidationError`` to specify a HTTP status code for the error response.
-* Errors are logged in ``handle_error``.
+* Improved error logging.
 * Add ``'query'`` as a valid target name.
 * Allow a list of validators to be passed to an ``Arg`` or ``Parser.parse``.
 * A more useful ``__repr__`` for ``Arg``.